requirements.txt
streamlit==1.26.0
requests==2.31.0
matplotlib==3.7.1
jinja2==3.1.2
python-dotenv==1.0.0
openai>=1.0.0